Revert "Support file chunking and add support for s2 coverings"
